test: add stock entry pcs validation and precision test case to workflow
